api/requirements.txt
anyio==3.6.1
asgiref==3.5.2
beautifulsoup4==4.11.1
blis==0.7.7
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==37.0.2
cymem==2.0.6
dnspython==2.2.1
email-validator==1.2.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.2.0/en_core_web_md-3.2.0-py3-none-any.whl
fastapi==0.79.1
Flask==2.1.1
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
langcodes==3.3.0
limits==1.6
lxml==4.9.1
MarkupSafe==2.1.1
murmurhash==1.0.7
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.1
orjson==3.7.12
packaging==21.3
pathy==0.6.1
preshed==3.0.6
protobuf==4.21.6
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
simplegmail==4.0.4
six==1.16.0
slowapi==0.1.6
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
starlette==0.19.1
thinc==8.0.15
tqdm==4.64.0
typer==0.4.1
typing-extensions==4.1.1
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
wasabi==0.9.1
watchgod==0.8.2
websockets==10.3
Werkzeug==2.1.1
zipp==3.8.0
